requirements.txt
six==1.15.0
Flask~=2.0.1
gunicorn==19.9.0
google-cloud-dialogflow==2.6.0
google-api-core==1.31.0
protobuf==3.17.3
